Fix up compilation so that Prosidy can be built sans-server
The server adds quite a bit of bloat to Prosidy, and for folks that
don't require it I'd rather make it optional.

Although I'm still just builting releases with the server, if you pull
the repo and build with `--no-default-features` you'll get a lighter
version of the executable without the server.